requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
databases==0.9.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
